requirements.txt
langchain==0.0.347
langchain-experimental==0.0.44
openai==1.3.7
botbuilder-integration-aiohttp>=4.14.4
python-dotenv
sqlalchemy<2.0.0
pyodbc
pandas
